test: qc-pay-api 테스트 코드 리팩토링  - kotest 테스트 코드로 변경  - Mock 생성 방식 변경

+++ qc-core/qc-pay/qc-pay-core/src/main/kotlin/com/eager/questioncloud/payment/domain/QuestionPaymentCoupon.kt
@@ -15,18 +15,18 @@ class QuestionPaymentCoupon(
 ) {
     fun calcDiscount(originalAmount: Int): Int {
         if (couponType == CouponType.Fixed) {
-            return max((originalAmount - value).toDouble(), 0.0).toInt()
+            return max(originalAmount - value, 0)
         }
-
+        
         if (couponType == CouponType.Percent) {
             val discountRate = value / 100.0
             val discountAmount = floor(originalAmount * discountRate).toInt()
             return originalAmount - discountAmount
         }
-
+        
         throw CoreException(Error.FAIL_USE_COUPON)
     }
-
+    
     companion object {
         fun create(userCouponId: Long, coupon: Coupon): QuestionPaymentCoupon {
             return QuestionPaymentCoupon(
